+++ .gitignore
@@ -1,3 +1,6 @@
+tests
+temp
+
 # ---> C
 # Prerequisites
 *.d
