--- exec_file
@@ -1,9 +1,0 @@
-.name "Simple"
-.comment "Let's get started"
-
-l2:             sti r1,%:live,%1
-                and r1,%0,r1
-
-live: live %1
-                zjmp %:live
-
