+++ include/asm.h
@@ -1,27 +1,28 @@
-#ifndef ASSEMBLER_H
-#define ASSEMBLER_H
+#ifndef _ASM_H
+#define _ASM_H
 
 #include <stdlib.h>
 #include <stdio.h>
 #include <fcntl.h>
 #include <unistd.h>
+#include <stdbool.h>
 #include <string.h> // replace with original implementations
 
+#include "../include/op.h"
+
+#define TAB_CHAR 9
+#define DOUBLE_QUOTE 34
+
 #define EXECUTABLE_EXT ".cor"
+#define DIRECTIVE_CHAR '.'
 
-// Taken from op.h -- Do I need header guards?
-#define COMMENT_CHAR '#'
-#define LABEL_CHAR ':'
-#define DIRECT_CHAR '%'
-#define SEPARATOR_CHAR ','
-
-typedef struct program {
+typedef struct source {
     const char *file_name;
-    int line_ct;
+    int line_num;
     FILE *fp;
     size_t buf;
     char *line;
-} prog;
+} src;
 
 typedef struct executable {
     char *file_name;
@@ -31,13 +32,14 @@ typedef struct executable {
 typedef struct token{
     int id;
     char *str;
-    int len;
+    size_t len;
     struct token *next;
 } token;
 
 typedef struct token_list{
     token *head;
     token *tail;
+    char *err_msg;
 } token_list;
 
 // not sure if this is needed
@@ -50,14 +52,6 @@ enum token_types{
     COMMENT, // begin with '#'
     SEPARATOR, // ','
     DIRECT, // begin with '%'
-    // Single chars
-    LEFT_PARENTHESIS,
-    RIGHT_PARENTHESIS,
-    LEFT_BRACKET,
-    RIGHT_BRACKET,
-    LEFT_BRACE,
-    RIGHT_BRACE,
-    EQUAL,
 };
 
 #endif
