+++ src/vm.c
@@ -1,0 +1,1 @@
+#include "../include/vm.h"
