+++ temp/op.c
@@ -1,4 +1,4 @@
-//#include "corewar.h"
+#include "../include/op.h"
 
 const op_t op_tab[] = {
   {"live", 1, {T_DIR}, 1, 10, inst_live},
