--- tests/asm_test_files/complex.s
@@ -1,15 +1,0 @@
-.name "Complex"
-.comment "Let's do it!"
-sti     r1,%:live,%1
-sti     r1,%:live2,%1
-ld      %1,r3
-ld      %33,r6
-#While (r2 < 10)
-forks:
-add     r2,r3,r2        #increment r2
-xor     r2,%15,r4       #if (r4) {carry = 0}
-live2:
-        live %4
-zjmp    %:endwhile      #if (carry)
-fork    %:forks
-ld      %0,r4           #carry = 1
